+++ packages/stream_chat_localizations/example/lib/add_new_lang.dart
@@ -383,6 +383,9 @@ class NnStreamChatLocalizations extends GlobalStreamChatLocalizations {
 
   @override
   String get replyToMessageLabel => 'Reply to Message';
+
+  @override
+  String get slowModeOnLabel => 'Slow mode ON';
 }
 
 void main() async {