+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_en.dart
@@ -359,4 +359,7 @@ class StreamChatLocalizationsEn extends GlobalStreamChatLocalizations {
 
   @override
   String get replyToMessageLabel => 'Reply to Message';
+
+  @override
+  String get slowModeOnLabel => 'Slow mode ON';
 }