+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_es.dart
@@ -364,4 +364,7 @@ class StreamChatLocalizationsEs extends GlobalStreamChatLocalizations {
 
   @override
   String get replyToMessageLabel => 'Responder al Mensaje';
+
+  @override
+  String get slowModeOnLabel => 'Modo lento activado';
 }