+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_fr.dart
@@ -363,4 +363,7 @@ class StreamChatLocalizationsFr extends GlobalStreamChatLocalizations {
 
   @override
   String get replyToMessageLabel => 'Répondre au Message';
+
+  @override
+  String get slowModeOnLabel => 'Mode lent activé';
 }